feat(frontend): 重构前端应用并添加新功能
- 重新组织前端代码结构，分离登录、注册和仪表盘页面
- 添加登录和注册页面组件
- 实现仪表盘页面，包含任务提交和状态检查功能
-集成 React Router 用于页面路由管理
- 优化样式和布局
