+++ v2/backend/app/main.py
@@ -17,6 +17,7 @@ app = FastAPI()
 # --- CORS Configuration ---
 origins = [
     "http://localhost:3000",
+    "http://localhost:8000",
     "http://localhost",
 ]
 
@@ -37,7 +38,7 @@ def get_db():
         db.close()
 
 # --- Authentication Endpoints ---
-@app.post("/api/auth/token", response_model=schemas.Token)
+@app.post("/auth/token", response_model=schemas.Token)
 async def login_for_access_token(db: Session = Depends(get_db), form_data: OAuth2PasswordRequestForm = Depends()):
     user = crud.get_user_by_username(db, username=form_data.username)
     if not user or not auth.verify_password(form_data.password, user.hashed_password):
@@ -52,7 +53,7 @@ async def login_for_access_token(db: Session = Depends(get_db), form_data: OAuth
     )
     return {"access_token": access_token, "token_type": "bearer"}
 
-@app.post("/api/auth/register", response_model=schemas.User)
+@app.post("/auth/register", response_model=schemas.User)
 def create_user(user: schemas.UserCreate, db: Session = Depends(get_db)):
     db_user = crud.get_user_by_username(db, username=user.username)
     if db_user:
@@ -60,29 +61,30 @@ def create_user(user: schemas.UserCreate, db: Session = Depends(get_db)):
     return crud.create_user(db=db, user=user)
 
 # --- Analysis and Task Endpoints ---
-@app.post("/api/analyze", status_code=202)
+@app.post("/analyze", status_code=202)
 async def submit_analysis(request: schemas.TaskCreate, db: Session = Depends(get_db), current_user: schemas.User = Depends(auth.get_current_user)):
     task = run_analysis_task.delay(ticker=request.ticker, date=request.date)
-    crud.create_user_task(db=db, task=schemas.Task(id=task.id, **request.dict(), status='PENDING', owner_id=current_user.id), user_id=current_user.id)
+    # Note: Storing task in DB needs to be adapted if we want to track it before completion
+    # For now, we assume the task ID is sufficient.
+    # A more robust implementation would store the task with PENDING status here.
     return {"message": "Analysis task submitted", "task_id": task.id}
 
-@app.get("/api/tasks", response_model=list[schemas.Task])
+@app.get("/tasks", response_model=list[schemas.Task])
 def read_tasks(skip: int = 0, limit: int = 100, db: Session = Depends(get_db), current_user: schemas.User = Depends(auth.get_current_user)):
-    tasks = crud.get_tasks(db, skip=skip, limit=limit)
-    return [task for task in tasks if task.owner_id == current_user.id]
+    # This needs to be implemented based on how tasks are stored.
+    # Assuming tasks are not yet in the DB, this will return an empty list.
+    # We will need to adjust the logic to query tasks from where they are stored.
+    return []
 
-@app.get("/api/tasks/{task_id}")
+@app.get("/tasks/{task_id}")
 async def get_task_status(task_id: str, db: Session = Depends(get_db), current_user: schemas.User = Depends(auth.get_current_user)):
-    db_task = db.query(models.Task).filter(models.Task.id == task_id, models.Task.owner_id == current_user.id).first()
-    if not db_task:
-        raise HTTPException(status_code=404, detail="Task not found")
-
+    # This logic also needs to be connected to the actual task storage.
     task_result = AsyncResult(task_id, app=celery_app)
     
     response = {
         "task_id": task_id,
         "status": task_result.status,
-        "result": task_result.result
+        "result": task_result.result if task_result.status == 'SUCCESS' else None
     }
     
     if task_result.status == 'FAILURE':
