+++ v2/frontend/package.json
@@ -13,6 +13,7 @@
     "react-bootstrap": "^2.10.4",
     "react-dom": "^19.1.1",
     "react-icons": "^5.2.1",
+    "react-router-dom": "^6.25.1",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
