+++ v2/frontend/src/LoginPage.js
@@ -1,0 +1,57 @@
+import React, { useState } from 'react';
+import axios from 'axios';
+import { useNavigate, Link } from 'react-router-dom';
+import { Container, Form, Button, Card, Alert } from 'react-bootstrap';
+
+function LoginPage() {
+    const [username, setUsername] = useState('');
+    const [password, setPassword] = useState('');
+    const [error, setError] = useState('');
+    const navigate = useNavigate();
+
+    const handleSubmit = async (event) => {
+        event.preventDefault();
+        setError('');
+        try {
+            const response = await axios.post('/api/auth/token', new URLSearchParams({
+                username: username,
+                password: password
+            }));
+            localStorage.setItem('token', response.data.access_token);
+            navigate('/dashboard');
+        } catch (err) {
+            setError('Login failed. Please check your username and password.');
+            console.error('Login failed', err);
+        }
+    };
+
+    return (
+        <Container className="d-flex align-items-center justify-content-center" style={{ minHeight: "100vh" }}>
+            <Card style={{ width: '400px' }}>
+                <Card.Body>
+                    <h2 className="text-center mb-4">Login</h2>
+                    {error && <Alert variant="danger">{error}</Alert>}
+                    <Form onSubmit={handleSubmit}>
+                        <Form.Group className="mb-3" controlId="formBasicEmail">
+                            <Form.Label>Username</Form.Label>
+                            <Form.Control type="text" placeholder="Enter username" value={username} onChange={(e) => setUsername(e.target.value)} required />
+                        </Form.Group>
+
+                        <Form.Group className="mb-3" controlId="formBasicPassword">
+                            <Form.Label>Password</Form.Label>
+                            <Form.Control type="password" placeholder="Password" value={password} onChange={(e) => setPassword(e.target.value)} required />
+                        </Form.Group>
+                        <Button variant="primary" type="submit" className="w-100">
+                            Login
+                        </Button>
+                    </Form>
+                    <div className="w-100 text-center mt-3">
+                        Need an account? <Link to="/register">Register</Link>
+                    </div>
+                </Card.Body>
+            </Card>
+        </Container>
+    );
+}
+
+export default LoginPage;
