+++ v2/frontend/src/RegisterPage.js
@@ -1,0 +1,53 @@
+import React, { useState } from 'react';
+import axios from 'axios';
+import { useNavigate, Link } from 'react-router-dom';
+import { Container, Form, Button, Card, Alert } from 'react-bootstrap';
+
+function RegisterPage() {
+    const [username, setUsername] = useState('');
+    const [password, setPassword] = useState('');
+    const [error, setError] = useState('');
+    const navigate = useNavigate();
+
+    const handleSubmit = async (event) => {
+        event.preventDefault();
+        setError('');
+        try {
+            await axios.post('/api/auth/register', { username, password });
+            navigate('/login');
+        } catch (err) {
+            setError('Registration failed. The username might already be taken.');
+            console.error('Registration failed', err);
+        }
+    };
+
+    return (
+        <Container className="d-flex align-items-center justify-content-center" style={{ minHeight: "100vh" }}>
+            <Card style={{ width: '400px' }}>
+                <Card.Body>
+                    <h2 className="text-center mb-4">Register</h2>
+                    {error && <Alert variant="danger">{error}</Alert>}
+                    <Form onSubmit={handleSubmit}>
+                        <Form.Group className="mb-3" controlId="formBasicEmail">
+                            <Form.Label>Username</Form.Label>
+                            <Form.Control type="text" placeholder="Choose a username" value={username} onChange={(e) => setUsername(e.target.value)} required />
+                        </Form.Group>
+
+                        <Form.Group className="mb-3" controlId="formBasicPassword">
+                            <Form.Label>Password</Form.Label>
+                            <Form.Control type="password" placeholder="Create a password" value={password} onChange={(e) => setPassword(e.target.value)} required />
+                        </Form.Group>
+                        <Button variant="primary" type="submit" className="w-100">
+                            Register
+                        </Button>
+                    </Form>
+                    <div className="w-100 text-center mt-3">
+                        Already have an account? <Link to="/login">Login</Link>
+                    </div>
+                </Card.Body>
+            </Card>
+        </Container>
+    );
+}
+
+export default RegisterPage;
